+++ lib/src/widgets/LenoreDatePicker.dart
@@ -2,27 +2,32 @@ import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 
 import 'package:lenore_ui/src/Helpers/Constants.dart';
+import '../../jalali_calendar/jalali_calendar.dart';
+
+enum CalendarType { Jalali, Gregorian }
 
 class LenoreDatePicker extends StatelessWidget {
-  final DateTime? value;
-  final Function onPressed;
-  final BuildContext? context;
+  final String? value;
+  final BuildContext context;
   final String? hint;
   final String? label;
   final IconData? icon;
   final DateType dateType;
   final String emptyText;
-  const LenoreDatePicker(
-      {Key? key,
-      required this.onPressed,
-      this.value,
-      this.context,
-      this.hint,
-      this.label,
-      this.icon,
-      this.dateType = DateType.DateAndTime,
-      this.emptyText = 'Choose ...'})
-      : super(key: key);
+  final Function(String?) onChange;
+  final CalendarType calendarType;
+  const LenoreDatePicker({
+    Key? key,
+    required this.context,
+    required this.onChange,
+    this.value,
+    this.hint,
+    this.label,
+    this.icon,
+    this.dateType = DateType.DateAndTime,
+    this.emptyText = 'Choose ...',
+    this.calendarType = CalendarType.Gregorian,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -36,8 +41,10 @@ class LenoreDatePicker extends StatelessWidget {
             : null,
       ),
       child: InkWell(
-        onTap: () {
-          onPressed();
+        onTap: () async {
+          pickDate(this.context).then((value) {
+            onChange(value);
+          });
         },
         child: InputDecorator(
           decoration: InputDecoration(
@@ -65,11 +72,7 @@ class LenoreDatePicker extends StatelessWidget {
           ),
           child: Container(
             child: Text(
-              value == null
-                  ? emptyText
-                  : (dateType == DateType.DateAndTime
-                      ? value.toString()
-                      : DateFormat('yyyy-MM-dd').format(value!)),
+              value ?? emptyText,
               style: value == null
                   ? Theme.of(context)
                       .textTheme
@@ -81,5 +84,42 @@ class LenoreDatePicker extends StatelessWidget {
         ),
       ),
     );
+  }
+
+  Future<String?> pickDate(BuildContext context) async {
+    if (calendarType == CalendarType.Gregorian) {
+      final DateTime? date = await showDatePicker(
+        context: context,
+        initialDate: DateTime.now(),
+        firstDate: DateTime(2015, 8),
+        lastDate: DateTime(2101),
+      );
+
+      if (date != null) {
+        if (dateType == DateType.DateAndTime) {
+          final time = await showTimePicker(
+              context: context, initialTime: TimeOfDay.now());
+          if (time != null) {
+            return DateFormat('yyyy-MM-dd').format(date) +
+                ' ' +
+                '${time.hour < 10 ? '0${time.hour}' : time.hour}:${time.minute < 10 ? '0${time.minute}' : time.minute}:00';
+          } else {
+            return DateFormat('yyyy-MM-dd').format(date);
+          }
+        } else {
+          return DateFormat('yyyy-MM-dd').format(date);
+        }
+      } else {
+        return null;
+      }
+    } else {
+      String? date =
+          await jalaliCalendarPicker(context: context, showTimePicker: true);
+      if (date != null) {
+        return date;
+      } else {
+        return null;
+      }
+    }
   }
 }